[Gold II] Title: 가장 긴 증가하는 부분 수열 3, Time: 184 ms, Memory: 12200 KB -BaekjoonHub

+++ 백준/Gold/12738. 가장 긴 증가하는 부분 수열 3/가장 긴 증가하는 부분 수열 3.cc
@@ -12,7 +12,7 @@ const int MAX = 1000001;
 
 int n;
 vector<int> dp;
-vector<LL> arr(1000001);
+vector<int> arr(1000001);
 
 int main(void) {
     FASTIO;
